refactor: extract Discord message and stream helpers into utils
